# elm-frontend/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "吃饱了吗点餐平台",
    "description": "Tianjin University Software Engineering Project",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/businesses/{id}": {
      "get": {
        "summary": "getBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "getBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "updateBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "deleteBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "patch": {
        "summary": "patchBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "patchBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "summary": "getOrderById",
        "deprecated": false,
        "description": "",
        "operationId": "getOrderById",
        "tags": ["管理订单"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResult"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/foods": {
      "get": {
        "summary": "getAllFoods",
        "deprecated": false,
        "description": "",
        "operationId": "getAllFoods",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "business",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultListFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "addFood",
        "deprecated": false,
        "description": "",
        "operationId": "addFood",
        "tags": ["管理商品"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "createTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "creator": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "updater": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "deleted": {
                    "type": "boolean"
                  },
                  "foodName": {
                    "type": "string"
                  },
                  "foodExplain": {
                    "type": "string"
                  },
                  "foodImg": {
                    "type": "string"
                  },
                  "foodPrice": {
                    "type": "number"
                  },
                  "business": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "createTime": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updateTime": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "creator": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "updater": {
                        "type": "integer",
                        "format": "int64"
                      },
                      "deleted": {
                        "type": "boolean"
                      },
                      "businessName": {
                        "type": "string"
                      },
                      "businessOwner": {
                        "$ref": "#/components/schemas/User"
                      },
                      "businessAddress": {
                        "type": "string"
                      },
                      "businessExplain": {
                        "type": "string"
                      },
                      "businessImg": {
                        "type": "string"
                      },
                      "orderTypeId": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "startPrice": {
                        "type": "number"
                      },
                      "deliveryPrice": {
                        "type": "number"
                      },
                      "remarks": {
                        "type": "string"
                      }
                    },
                    "required": ["id"]
                  },
                  "remarks": {
                    "type": "string"
                  }
                },
                "required": ["foodName", "foodPrice", "business"]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/foods/{id}": {
      "get": {
        "summary": "返回查询到的一条商品记录",
        "deprecated": false,
        "description": "",
        "operationId": "getFoodById",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新商品信息",
        "deprecated": false,
        "description": "更新指定ID的商品信息",
        "operationId": "updateFood",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Food"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除商品",
        "deprecated": false,
        "description": "删除指定ID的商品",
        "operationId": "deleteFood",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Authority": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Business": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "businessName": {
            "type": "string"
          },
          "businessOwner": {
            "$ref": "#/components/schemas/User"
          },
          "businessAddress": {
            "type": "string"
          },
          "businessExplain": {
            "type": "string"
          },
          "businessImg": {
            "type": "string"
          },
          "orderTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "startPrice": {
            "type": "number"
          },
          "deliveryPrice": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
...
      "User": {
        "required": ["username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "authorities": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Authority"
            }
          }
        }
      },
      "HttpResultBusiness": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Business"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Food": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "foodName": {
            "type": "string"
          },
          "foodExplain": {
            "type": "string"
          },
          "foodImg": {
            "type": "string"
          },
          "foodPrice": {
            "type": "number"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "remarks": {
            "type": "string"
          }
        }
      },
      "HttpResultFood": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Food"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "HttpResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {}
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "HttpResultListFood": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Food"
            }
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}